changed multiplier factor for cam frame size from int to float for finer adjustment

+++ src/quad.h
@@ -24,8 +24,8 @@ public:
     ofTexture camTexture;
     int camWidth;
     int camHeight;
-    int camMultX;
-    int camMultY;
+    float camMultX;
+    float camMultY;
 
     int quadNumber;
